requirements.txt
Flask==3.0.3
peewee==3.17.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
paho-mqtt==2.1.0
